+++ pkg/services/baremetal/baremetal/baremetal_test.go
@@ -414,161 +414,6 @@ var _ = Describe("Test consumerRefMatches", func() {
 	)
 })
 
-var _ = Describe("Test findOwnerRefFromList", func() {
-	type testCaseFindOwnerRefFromList struct {
-		RefList          []metav1.OwnerReference
-		ExpectedPosition *int
-	}
-
-	objectMeta := metav1.ObjectMeta{
-		Name:      "bm-machine",
-		Namespace: "default",
-	}
-	objectType := metav1.TypeMeta{
-		Kind:       "HetznerBareMetalMachine",
-		APIVersion: "v1beta1",
-	}
-
-	DescribeTable("Test findOwnerRefFromList",
-		func(tc testCaseFindOwnerRefFromList) {
-			position, found := findOwnerRefFromList(tc.RefList, objectType, objectMeta)
-
-			if tc.ExpectedPosition != nil {
-				Expect(found).To(BeTrue())
-				Expect(position).To(Equal(*tc.ExpectedPosition))
-			} else {
-				Expect(found).To(BeFalse())
-			}
-		},
-		Entry("Matching consumer", testCaseFindOwnerRefFromList{
-			RefList: []metav1.OwnerReference{
-				{
-					Name:       "bm-machine",
-					Kind:       "HetznerBareMetalMachine",
-					APIVersion: "v1beta1",
-				},
-			},
-			ExpectedPosition: pointer.Int(0),
-		}),
-		Entry("Matching consumer position 1", testCaseFindOwnerRefFromList{
-			RefList: []metav1.OwnerReference{
-				{
-					Name:       "bm-machine2",
-					Kind:       "HetznerBareMetalMachine",
-					APIVersion: "v1beta1",
-				},
-				{
-					Name:       "bm-machine",
-					Kind:       "HetznerBareMetalMachine",
-					APIVersion: "v1beta1",
-				},
-			},
-			ExpectedPosition: pointer.Int(1),
-		}),
-		Entry("Matching consumer position 1a", testCaseFindOwnerRefFromList{
-			RefList: []metav1.OwnerReference{
-				{
-					Name:       "bm-machine",
-					Kind:       "OtherBareMetalMachine",
-					APIVersion: "v1beta1",
-				},
-				{
-					Name:       "bm-machine",
-					Kind:       "HetznerBareMetalMachine",
-					APIVersion: "v1beta1",
-				},
-			},
-			ExpectedPosition: pointer.Int(1),
-		}),
-		Entry("Matching consumer position 1b", testCaseFindOwnerRefFromList{
-			RefList: []metav1.OwnerReference{
-				{
-					Name:       "bm-machine",
-					Kind:       "HetznerBareMetalMachine",
-					APIVersion: "hetzner/v1beta1",
-				},
-				{
-					Name:       "bm-machine",
-					Kind:       "HetznerBareMetalMachine",
-					APIVersion: "v1beta1",
-				},
-			},
-			ExpectedPosition: pointer.Int(1),
-		}),
-	)
-})
-
-var _ = Describe("Test removeOwnerRefFromList", func() {
-	type testCaseRemoveOwnerRefFromList struct {
-		RefList         []metav1.OwnerReference
-		ExpectedRefList []metav1.OwnerReference
-	}
-
-	objectMeta := metav1.ObjectMeta{
-		Name:      "bm-machine",
-		Namespace: "default",
-	}
-	objectType := metav1.TypeMeta{
-		Kind:       "HetznerBareMetalMachine",
-		APIVersion: "v1beta1",
-	}
-
-	expectedRefList3 := make([]metav1.OwnerReference, 0, 2)
-	expectedRefList3 = append(expectedRefList3, metav1.OwnerReference{
-		Name:       "bm-machine2",
-		Kind:       "HetznerBareMetalMachine",
-		APIVersion: "v1beta1",
-	})
-
-	DescribeTable("Test removeOwnerRefFromList",
-		func(tc testCaseRemoveOwnerRefFromList) {
-			refList := removeOwnerRefFromList(tc.RefList, objectType, objectMeta)
-			Expect(refList).To(Equal(tc.ExpectedRefList))
-		},
-		Entry("List of one matching entry", testCaseRemoveOwnerRefFromList{
-			RefList: []metav1.OwnerReference{
-				{
-					Name:       "bm-machine",
-					Kind:       "HetznerBareMetalMachine",
-					APIVersion: "v1beta1",
-				},
-			},
-			ExpectedRefList: []metav1.OwnerReference{},
-		}),
-		Entry("List of one non-matching entry", testCaseRemoveOwnerRefFromList{
-			RefList: []metav1.OwnerReference{
-				{
-					Name:       "bm-machine2",
-					Kind:       "HetznerBareMetalMachine",
-					APIVersion: "v1beta1",
-				},
-			},
-			ExpectedRefList: []metav1.OwnerReference{
-				{
-					Name:       "bm-machine2",
-					Kind:       "HetznerBareMetalMachine",
-					APIVersion: "v1beta1",
-				},
-			},
-		}),
-		Entry("Two entries with matching", testCaseRemoveOwnerRefFromList{
-			RefList: []metav1.OwnerReference{
-				{
-					Name:       "bm-machine2",
-					Kind:       "HetznerBareMetalMachine",
-					APIVersion: "v1beta1",
-				},
-				{
-					Name:       "bm-machine",
-					Kind:       "HetznerBareMetalMachine",
-					APIVersion: "v1beta1",
-				},
-			},
-			ExpectedRefList: expectedRefList3,
-		}),
-	)
-})
-
 var _ = Describe("Test setOwnerRefInList", func() {
 	type testCaseSetOwnerRefInList struct {
 		RefList         []metav1.OwnerReference
