+++ pkg/utils/utils_test.go
@@ -20,6 +20,8 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"github.com/syself/cluster-api-provider-hetzner/pkg/utils"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/utils/pointer"
 )
 
 var _ = DescribeTable("LabelsToLabelSelector",
@@ -112,4 +114,149 @@ var _ = Describe("FilterStringFromList", func() {
 		},
 		Entry("entry1", []string{"a", "b", "c"}, "a", []string{"b", "c"}),
 		Entry("entry2", []string{"a", "b", "c"}, "d", []string{"a", "b", "c"}))
+})
+
+var _ = Describe("Test removeOwnerRefFromList", func() {
+	type testCaseRemoveOwnerRefFromList struct {
+		RefList         []metav1.OwnerReference
+		ExpectedRefList []metav1.OwnerReference
+	}
+
+	name := "bm-machine"
+	kind := "HetznerBareMetalMachine"
+	apiVersion := "v1beta1"
+
+	expectedRefList3 := make([]metav1.OwnerReference, 0, 2)
+	expectedRefList3 = append(expectedRefList3, metav1.OwnerReference{
+		Name:       "bm-machine2",
+		Kind:       "HetznerBareMetalMachine",
+		APIVersion: "v1beta1",
+	})
+
+	DescribeTable("Test RemoveOwnerRefFromList",
+		func(tc testCaseRemoveOwnerRefFromList) {
+			refList := utils.RemoveOwnerRefFromList(tc.RefList, name, kind, apiVersion)
+			Expect(refList).To(Equal(tc.ExpectedRefList))
+		},
+		Entry("List of one matching entry", testCaseRemoveOwnerRefFromList{
+			RefList: []metav1.OwnerReference{
+				{
+					Name:       "bm-machine",
+					Kind:       "HetznerBareMetalMachine",
+					APIVersion: "v1beta1",
+				},
+			},
+			ExpectedRefList: []metav1.OwnerReference{},
+		}),
+		Entry("List of one non-matching entry", testCaseRemoveOwnerRefFromList{
+			RefList: []metav1.OwnerReference{
+				{
+					Name:       "bm-machine2",
+					Kind:       "HetznerBareMetalMachine",
+					APIVersion: "v1beta1",
+				},
+			},
+			ExpectedRefList: []metav1.OwnerReference{
+				{
+					Name:       "bm-machine2",
+					Kind:       "HetznerBareMetalMachine",
+					APIVersion: "v1beta1",
+				},
+			},
+		}),
+		Entry("Two entries with matching", testCaseRemoveOwnerRefFromList{
+			RefList: []metav1.OwnerReference{
+				{
+					Name:       "bm-machine2",
+					Kind:       "HetznerBareMetalMachine",
+					APIVersion: "v1beta1",
+				},
+				{
+					Name:       "bm-machine",
+					Kind:       "HetznerBareMetalMachine",
+					APIVersion: "v1beta1",
+				},
+			},
+			ExpectedRefList: expectedRefList3,
+		}),
+	)
+})
+
+var _ = Describe("Test FindOwnerRefFromList", func() {
+	type testCaseFindOwnerRefFromList struct {
+		RefList          []metav1.OwnerReference
+		ExpectedPosition *int
+	}
+
+	name := "bm-machine"
+	kind := "HetznerBareMetalMachine"
+	apiVersion := "v1beta1"
+
+	DescribeTable("Test FindOwnerRefFromList",
+		func(tc testCaseFindOwnerRefFromList) {
+			position, found := utils.FindOwnerRefFromList(tc.RefList, name, kind, apiVersion)
+
+			if tc.ExpectedPosition != nil {
+				Expect(found).To(BeTrue())
+				Expect(position).To(Equal(*tc.ExpectedPosition))
+			} else {
+				Expect(found).To(BeFalse())
+			}
+		},
+		Entry("Matching consumer", testCaseFindOwnerRefFromList{
+			RefList: []metav1.OwnerReference{
+				{
+					Name:       "bm-machine",
+					Kind:       "HetznerBareMetalMachine",
+					APIVersion: "v1beta1",
+				},
+			},
+			ExpectedPosition: pointer.Int(0),
+		}),
+		Entry("Matching consumer position 1", testCaseFindOwnerRefFromList{
+			RefList: []metav1.OwnerReference{
+				{
+					Name:       "bm-machine2",
+					Kind:       "HetznerBareMetalMachine",
+					APIVersion: "v1beta1",
+				},
+				{
+					Name:       "bm-machine",
+					Kind:       "HetznerBareMetalMachine",
+					APIVersion: "v1beta1",
+				},
+			},
+			ExpectedPosition: pointer.Int(1),
+		}),
+		Entry("Matching consumer position 1a", testCaseFindOwnerRefFromList{
+			RefList: []metav1.OwnerReference{
+				{
+					Name:       "bm-machine",
+					Kind:       "OtherBareMetalMachine",
+					APIVersion: "v1beta1",
+				},
+				{
+					Name:       "bm-machine",
+					Kind:       "HetznerBareMetalMachine",
+					APIVersion: "v1beta1",
+				},
+			},
+			ExpectedPosition: pointer.Int(1),
+		}),
+		Entry("Matching consumer position 1b", testCaseFindOwnerRefFromList{
+			RefList: []metav1.OwnerReference{
+				{
+					Name:       "bm-machine",
+					Kind:       "HetznerBareMetalMachine",
+					APIVersion: "hetzner/v1beta1",
+				},
+				{
+					Name:       "bm-machine",
+					Kind:       "HetznerBareMetalMachine",
+					APIVersion: "v1beta1",
+				},
+			},
+			ExpectedPosition: pointer.Int(1),
+		}),
+	)
 })
